refactor: move frame animation to sprite
