+++ packages/effects-core/src/components/base-render-component.ts
@@ -8,12 +8,11 @@ import type { Maskable } from '../material';
 import {
   MaskMode, MaskProcessor, Material, getPreMultiAlpha, setBlendMode, setMaskMode, setSideMode,
 } from '../material';
-import type { BoundingBoxTriangle, HitTestTriangleParams, splitsDataType } from '../plugins';
+import type { BoundingBoxTriangle, HitTestTriangleParams } from '../plugins';
 import { MeshCollider } from '../plugins';
 import type { Renderer } from '../render';
 import { Geometry } from '../render';
 import { itemFrag, itemVert } from '../shader';
-import { rotateVec2 } from '../shape';
 import { Texture } from '../texture';
 import { RendererComponent } from './renderer-component';
 
@@ -26,46 +25,29 @@ export interface ItemRenderer extends Required<Omit<spec.RendererOptions, 'textu
 }
 
 // TODO: Add to spec
-interface BaseRenderComponentData extends spec.ComponentData {
+interface MaskableGraphicData extends spec.ComponentData {
   renderer: spec.RendererOptions,
-  /**
-   * added by loader
-   * @default null
-   */
-  splits?: spec.SplitParameter[],
-  /**
-   * 图层元素贴图变化属性
-   */
-  textureSheetAnimation?: spec.TextureSheetAnimation,
-  geometry?: spec.DataPath,
   mask?: spec.MaskOptions,
 }
-
-const singleSplits: splitsDataType = [[0, 0, 1, 1, 0]];
 
 /**
  * @since 2.1.0
  */
-export class BaseRenderComponent extends RendererComponent implements Maskable {
+export class MaskableGraphic extends RendererComponent implements Maskable {
   interaction?: { behavior: spec.InteractBehavior };
   renderer: ItemRenderer;
   geometry: Geometry;
-  /**
-   * @internal
-  */
-  splits: splitsDataType = singleSplits;
 
   protected visible = true;
-  protected textureSheetAnimation?: spec.TextureSheetAnimation;
   protected readonly maskManager: MaskProcessor;
 
   /**
    * 用于点击测试的碰撞器
    */
   protected meshCollider = new MeshCollider();
+  protected defaultGeometry: Geometry;
 
   private _color = new Color(1, 1, 1, 1);
-  private defaultGeometry: Geometry;
 
   /**
    *
@@ -261,99 +243,6 @@ export class BaseRenderComponent extends RendererComponent implements Maskable {
     return boundingBox;
   }
 
-  private getItemGeometryData (renderer: ItemRenderer) {
-    const { splits, textureSheetAnimation } = this;
-    const sx = 1, sy = 1;
-    const geometry = this.defaultGeometry;
-
-    const originData = [-.5, .5, -.5, -.5, .5, .5, .5, -.5];
-    const aUV = [];
-    const index = [];
-    const position = [];
-
-    if (splits.length === 1) {
-      const split: number[] = textureSheetAnimation ? [0, 0, 1, 1, splits[0][4] as number] : splits[0] as number[];
-      const texOffset = split[4] ? [0, 0, 1, 0, 0, 1, 1, 1] : [0, 1, 0, 0, 1, 1, 1, 0];
-      const tox = split[0];
-      const toy = split[1];
-      const tsx = split[4] ? split[3] : split[2];
-      const tsy = split[4] ? split[2] : split[3];
-
-      aUV.push(
-        texOffset[0] * tsx + tox, texOffset[1] * tsy + toy,
-        texOffset[2] * tsx + tox, texOffset[3] * tsy + toy,
-        texOffset[4] * tsx + tox, texOffset[5] * tsy + toy,
-        texOffset[6] * tsx + tox, texOffset[7] * tsy + toy,
-      );
-      position.push(
-        originData[0], originData[1], 0.0,
-        originData[2], originData[3], 0.0,
-        originData[4], originData[5], 0.0,
-        originData[6], originData[7], 0.0
-      );
-      index.push(0, 1, 2, 2, 1, 3);
-    } else {
-      // TODO: 原有打包纹理拆分逻辑，待移除
-      //-------------------------------------------------------------------------
-
-      const col = 2;
-      const row = 2;
-
-      for (let x = 0; x < col; x++) {
-        for (let y = 0; y < row; y++) {
-          const base = (y * 2 + x) * 4;
-          // @ts-expect-error
-          const split: number[] = textureSheetAnimation ? [0, 0, 1, 1, splits[0][4]] : splits[y * 2 + x];
-          const texOffset = split[4] ? [0, 0, 1, 0, 0, 1, 1, 1] : [0, 1, 0, 0, 1, 1, 1, 0];
-          const dw = ((x + x + 1) / col - 1) / 2;
-          const dh = ((y + y + 1) / row - 1) / 2;
-          const tox = split[0];
-          const toy = split[1];
-          const tsx = split[4] ? split[3] : split[2];
-          const tsy = split[4] ? split[2] : split[3];
-          const origin = [
-            originData[0] / col + dw,
-            originData[1] / row + dh,
-            originData[2] / col + dw,
-            originData[3] / row + dh,
-            originData[4] / col + dw,
-            originData[5] / row + dh,
-            originData[6] / col + dw,
-            originData[7] / row + dh,
-          ];
-
-          aUV.push(
-            texOffset[0] * tsx + tox, texOffset[1] * tsy + toy,
-            texOffset[2] * tsx + tox, texOffset[3] * tsy + toy,
-            texOffset[4] * tsx + tox, texOffset[5] * tsy + toy,
-            texOffset[6] * tsx + tox, texOffset[7] * tsy + toy,
-          );
-          position.push((origin[0]) * sx, (origin[1]) * sy, 0.0,
-            (origin[2]) * sx, (origin[3]) * sy, 0.0,
-            (origin[4]) * sx, (origin[5]) * sy, 0.0,
-            (origin[6]) * sx, (origin[7]) * sy, 0.0);
-          index.push(base, 1 + base, 2 + base, 2 + base, 1 + base, 3 + base);
-        }
-      }
-    }
-
-    geometry.setAttributeData('aPos', new Float32Array(position));
-
-    return { index, aUV };
-  }
-
-  private configureDefaultGeometry (renderer: ItemRenderer) {
-    const geoData = this.getItemGeometryData(renderer);
-    const { index, aUV } = geoData;
-    const geometry = this.defaultGeometry;
-
-    geometry.setIndexData(new Uint16Array(index));
-    geometry.setAttributeData('aUV', new Float32Array(aUV));
-    geometry.setDrawCount(index.length);
-
-    return geometry;
-  }
-
   private configureMaterial (renderer: ItemRenderer): Material {
     const { side, occlusion, blending: blendMode, mask, texture } = renderer;
     const maskMode = this.maskManager.maskMode;
@@ -411,19 +300,14 @@ export class BaseRenderComponent extends RendererComponent implements Maskable {
   override fromData (data: unknown): void {
     super.fromData(data);
 
-    const baseRenderComponentData = (data as BaseRenderComponentData);
+    const baseRenderComponentData = (data as MaskableGraphicData);
     const renderer = baseRenderComponentData.renderer ?? {};
-    const splits = baseRenderComponentData.splits;
-    const textureSheetAnimation = baseRenderComponentData.textureSheetAnimation;
 
     const maskOptions = baseRenderComponentData.mask;
 
     if (maskOptions) {
       this.maskManager.setMaskOptions(maskOptions);
     }
-
-    this.splits = splits || singleSplits;
-    this.textureSheetAnimation = textureSheetAnimation;
 
     this.renderer = {
       renderMode: renderer.renderMode ?? spec.RenderMode.MESH,
@@ -436,58 +320,5 @@ export class BaseRenderComponent extends RendererComponent implements Maskable {
     };
 
     this.configureMaterial(this.renderer);
-
-    if (baseRenderComponentData.geometry) {
-      const baseGeometry = this.engine.findObject<Geometry>(baseRenderComponentData.geometry);
-
-      const uvTransform = baseRenderComponentData.splits && !baseRenderComponentData.textureSheetAnimation ? baseRenderComponentData.splits[0] : singleSplits[0];
-      const x = uvTransform[0];
-      const y = uvTransform[1];
-      const isRotate90 = Boolean(uvTransform[4]);
-      const width = isRotate90 ? uvTransform[3] : uvTransform[2];
-      const height = isRotate90 ? uvTransform[2] : uvTransform[3];
-      const angle = isRotate90 ? -Math.PI / 2 : 0;
-
-      const aUV = baseGeometry.getAttributeData('aUV');
-      const aPos = baseGeometry.getAttributeData('aPos');
-      const indices = baseGeometry.getIndexData();
-
-      const tempPosition: spec.vec2 = [0, 0];
-
-      if (aUV && aPos && indices) {
-        const vertexCount = aUV.length / 2;
-
-        for (let i = 0; i < vertexCount; i++) {
-          const positionOffset = i * 3;
-          const uvOffset = i * 2;
-          const positionX = aPos[positionOffset];
-          const positionY = aPos[positionOffset + 1];
-
-          tempPosition[0] = positionX ;
-          tempPosition[1] = positionY ;
-          rotateVec2(tempPosition, tempPosition, angle);
-
-          aUV[uvOffset] = (tempPosition[0] + 0.5) * width + x;
-          aUV[uvOffset + 1] = (tempPosition[1] + 0.5) * height + y;
-        }
-
-        this.geometry.setAttributeData('aPos', aPos.slice());
-        this.geometry.setAttributeData('aUV', aUV.slice());
-        this.geometry.setIndexData(indices.slice());
-        this.geometry.setDrawCount(indices.length);
-      }
-
-      this.geometry.subMeshes.length = 0;
-      for (const subMesh of baseGeometry.subMeshes) {
-        this.geometry.subMeshes.push({
-          offset: subMesh.offset,
-          indexCount:  subMesh.indexCount,
-          vertexCount:  subMesh.vertexCount,
-        });
-      }
-    } else {
-      this.geometry = this.defaultGeometry;
-      this.configureDefaultGeometry(this.renderer);
-    }
   }
 }
